Remove documentation from website repo
We were previously commiting the rendered output of sphinx builds from
the documentation repo into the website repo on each change made to the
Qiskit/qiskit repo. There is no reason to do this, as it just balloons
the size of the repository making it hard to work with and needlessly
increases the amount of times we have to copy easily generatable
static artifacts around. Unfortunately the damage to this repo is already
done because the large copies of all the binary static artifacts are now
in the git history forever. But, to address this problem in the future
the docs jobs in Qiskit/qiskit are changing to just upload directly
cloud object store where the site is hosted now, which both makes
updating documentation and the website faster, reduces complexity of the
deployment system, and makes everything more reliable. To ensure that
the qiskit.org deploys no longer try to overwrite the deployed
documentation this commit removes the local copy of the documentation
and updates the rclone command in the deploy script to not touch the
documentation content.